Delete gradlew to test if it fixes the build
The gradle wrapper included in this repo is almost a year old - it's a major release away from what the new versions of GradleRIO use. Travis CI keeps using the wrapper instead of the updated version of Gradle included in the worker images. This just deletes gradlew to test if the build succeeds after.